requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.9
albumentations>=1.4.3
opencv-python>=4.9.0
numpy>=1.26
pandas>=2.2.2
matplotlib>=3.8
pyyaml>=6.0.1
tqdm>=4.66
pytorch-grad-cam>=1.5.2
